fixed main branch linter issues
